third_party/blink/web_tests/http/tests/security/referrer-on-client-redirect-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/resources/referrer-on-client-redirect.html' attempted to load subresource at 'https://127.0.0.1:8443/security/resources/referrer-on-client-redirect.html?frame' over broken HTTPS.
PASS: Referrer is empty
